Scan widget/controller/manager generalization
Generalize scan widget/controller/managers generation by splitting up into different types (base/monalisa/pointscan for now), and setting which type should be loaded in the setup json file.
