+++ imswitch/imcontrol/model/managers/__init__.py
@@ -9,3 +9,4 @@ from .RecordingManager import RecordingManager, RecMode, SaveMode, SaveFormat
 from .SLMManager import SLMManager
 from .ScanManagerPointScan import ScanManagerPointScan
 from .ScanManagerBase import ScanManagerBase
+from .ScanManagerMoNaLISA import ScanManagerMoNaLISA
